Add files via upload
Complete Project zip file 


Note: Need to fix some major bugs, particularly in the writing state function. The data was mismatched; we need to fix 

bug was if i give 10 as a input means data[0] = '1' and data[1] = '0' we need to convert this into the proper data. i will fix ASAP